requirements.txt
xlsxwriter==3.0.3
colorama==0.4.3
camelot-py==0.10.1
pyfiglet==0.8.post1